refactor(misconf): remove unused methods from providers
