+++ pkg/iac/providers/aws/sns/sns.go
@@ -8,17 +8,6 @@ type SNS struct {
 	Topics []Topic
 }
 
-func NewTopic(arn string, metadata iacTypes.Metadata) *Topic {
-	return &Topic{
-		Metadata: metadata,
-		ARN:      iacTypes.String(arn, metadata),
-		Encryption: Encryption{
-			Metadata: metadata,
-			KMSKeyID: iacTypes.StringDefault("", metadata),
-		},
-	}
-}
-
 type Topic struct {
 	Metadata   iacTypes.Metadata
 	ARN        iacTypes.StringValue
