+++ pkg/iac/providers/github/branch_protections.go
@@ -8,7 +8,3 @@ type BranchProtection struct {
 	Metadata             iacTypes.Metadata
 	RequireSignedCommits iacTypes.BoolValue
 }
-
-func (b BranchProtection) RequiresSignedCommits() bool {
-	return b.RequireSignedCommits.IsTrue()
-}
